requirements.txt
flask==2.3.3
twilio==8.10.0
google-generativeai==0.3.0
python-dotenv==1.0.0
gunicorn==21.2.0